refactor: move useReviewsData to custom hook
